Modulo 1 APIs de IA Generativa/requirements.txt
# APIs de IA Generativa
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
ollama>=0.1.0

# Webapp ReactPy
reactpy>=1.0.0
uvicorn>=0.27.0

# Utilidades
pyyaml>=6.0
python-dotenv>=1.0.0
